split out the interface file that is getting too large
